fix(bugs): carte de mini-boss perdue une fois sur deux + spam d animation
Deux bugs releves dans le journal d une session de test jouee, sans rapport
avec le reequilibrage en cours.

1. LevelUpSystem.BuildWeaponCards : (int)GD.Randi() % (i + 1). Le cast d un
   uint PRECEDE le modulo -> index negatif une fois sur deux, donc
   ArgumentOutOfRangeException dans le shuffle. L exception etant avalee par
   le callback Godot, ShowWeaponDrop n affichait aucune carte : la recompense
   de mini-boss etait perdue une fois sur deux, SANS AUCUN SIGNE A L ECRAN.
   9 occurrences sur une seule session. Corrige en faisant le modulo en uint,
   comme LevelUpScreen et LevelSelectScreen le font deja.

2. There is no animation with name attack (144 occurrences). Les scenes
   archetype sont partagees par toute la faune data-driven, et les 5 golems
   slow_hunter n exposent que idle/move/death. Nouveau EnemyBase.PlayAnim :
   ne joue que si l animation existe, et RENVOIE si elle a demarre -- le cas
   death en depend, son QueueFree passant par AnimationFinished (un sprite
   sans death aurait laisse l ennemi mort a l ecran indefiniment).
   Applique aux 4 scenes archetype + RustedCore.

Au passage, calibrage du boss cale sur la SESSION JOUEE : maxHp 4000 -> 5000.
Le build joue fait 488 DPS la ou le calcul analytique projetait 310 (-40 %),
d ou un TTK de 18,7 s sous la fenetre 20-30 s. A 5000 : 25,6 s au banc de
non-regression, ~23 s pour le build joue. Ce boss ne se calibre que sur un
TTK joue (GDD 20.6).

Banc de non-regression : 0 erreur (contre 153 avant), TTK 25,6 s.
Pieges : PITFALLS sections Aleatoire et Animations d ennemis.

+++ src/Systems/LevelUpSystem.cs
@@ -368,11 +368,16 @@ public partial class LevelUpSystem : Node
             }
         }
 
-        // Shuffle et prendre count cartes
+        // Shuffle et prendre count cartes.
+        // ⚠ Le modulo se fait en UINT avant le cast : `(int)GD.Randi()` est négatif une fois sur
+        // deux (Randi() couvre tout l'espace uint), et `négatif % n` reste négatif en C# → index
+        // hors bornes. La récompense de mini-boss était perdue une fois sur deux, silencieusement :
+        // l'exception remontait dans le callback Godot, ShowWeaponDrop n'affichait aucune carte et
+        // seul le journal en gardait la trace.
         var shuffled = new System.Collections.Generic.List<LevelUpCardData>(available);
         for (int i = shuffled.Count - 1; i > 0; i--)
         {
-            int j = (int)GD.Randi() % (i + 1);
+            int j = (int)(GD.Randi() % (uint)(i + 1));
             (shuffled[i], shuffled[j]) = (shuffled[j], shuffled[i]);
         }
